Move CfgVehicles to dedicated .hpp file
